Pull the getCrumbsByColumn() function out
